pyproject.toml
[project]
name = "buja"
version = "0.1.0"
description = "BUJA 프로젝트 - Streamlit 기반"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "pyarrow>=14.0.0,<19.0.0",  # Windows build fix
    # Database
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",  # SQLite async driver (기본, Docker 불필요)
    # "asyncpg>=0.29.0",  # PostgreSQL async driver (선택사항) - Windows build issue
    "alembic>=1.13.0",
    # Cache
    "redis>=5.0.0",
    # Settings & Validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # LLM APIs
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    # HTTP Client
    "httpx>=0.25.0",
    # Security
    "bcrypt>=4.1.0",
    "cryptography>=41.0.0",
    # Environment
    "python-dotenv>=1.0.0",
    # Optimization
    "scipy>=1.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.7.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "pre-commit>=3.5.0",
    "playwright>=1.40.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true



